Refactor Users to the TS
