--- src/resources/users/user.model.js
@@ -1,20 +1,0 @@
-class User {
-  constructor({
-    id,
-    name,
-    login,
-    password
-  } = {}) {
-    this.id = id;
-    this.name = name;
-    this.login = login;
-    this.password = password;
-  }
-
-  static toResponse(user) {
-    const { id, name, login } = user;
-    return { id, name, login };
-  }
-}
-
-module.exports = User;
